Fixes for GPX handling.
* Double precision to avoid rounding errors on load.
* Predictable <time> tag in the GPX header.
* Updated and corrected reference files.

+++ Glint/SharedCode/JBGPXReader.h
@@ -20,7 +20,7 @@ typedef enum {
 	JBLocationMath *locationMath;
 
 	NSDate *lastReadDate;
-	float lastReadLat, lastReadLon, lastReadElevation;
+	double lastReadLat, lastReadLon, lastReadElevation;
 	CurrentlyReadingEnum currentlyReading;
 	BOOL shouldAddBreakMarker;
 }
